# services/rest-api/docs/swagger.yaml
basePath: /
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid email format
        type: string
    type: object
  handlers.SignInRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Password123!
        type: string
    type: object
  handlers.SignInResponse:
    properties:
      token:
        example: temporary_token
        type: string
    type: object
  handlers.SignUpRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Password123!
        type: string
    type: object
  handlers.SignUpResponse:
    properties:
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  handlers.ValidateTokenResponse:
    properties:
      user_id:
        example: "123"
        type: string
      valid:
        example: true
        type: boolean
    type: object
host: 88.218.169.245:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    REST API Gateway для микросервисной архитектуры на Go с gRPC

    Этот API предоставляет endpoints для аутентификации пользователей. Backend построен на микросервисной архитектуре с использованием gRPC для межсервисного взаимодействия.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Golang Microservices API
  version: "1.0"
paths:
  /api/v1/auth/signin:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя и получение токена
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход, токен выдан
          schema:
            $ref: '#/definitions/handlers.SignInResponse'
        "400":
          description: Невалидные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Неверный пароль
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с email и паролем
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/handlers.SignUpResponse'
        "400":
          description: Невалидные данные (email или пароль)
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /api/v1/auth/validate:
    get:
      consumes:
      - application/json
      description: Валидация JWT токена и получение информации о пользователе
      parameters:
      - default: Bearer temporary_token
        description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Токен валиден
          schema:
            $ref: '#/definitions/handlers.ValidateTokenResponse'
        "401":
          description: Отсутствует или невалидный токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Проверка токена
      tags:
      - auth
  /health:
    get:
      description: Проверка работоспособности сервиса
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health Check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"